requirements.txt
transformers
torch
opencv-python
opencv-python-headless
timm
sentence-transformers
faiss-cpu
fastapi
pydantic
uvicorn
python-multipart
Pillow
numpy
inference-sdk